+++ src/libs/DB.ts
@@ -1,13 +1,15 @@
 import path from 'node:path';
 
 import { PGlite } from '@electric-sql/pglite';
-import { drizzle as drizzlePg } from 'drizzle-orm/node-postgres';
-import { migrate as migratePg } from 'drizzle-orm/node-postgres/migrator';
+// import { drizzle as drizzlePg } from 'drizzle-orm/node-postgres';
+// import { migrate as migratePg } from 'drizzle-orm/node-postgres/migrator';
+import { Pool } from '@neondatabase/serverless';
+import { drizzle as drizzleNeon } from 'drizzle-orm/neon-serverless';
 import { drizzle as drizzlePglite, type PgliteDatabase } from 'drizzle-orm/pglite';
 import { migrate as migratePglite } from 'drizzle-orm/pglite/migrator';
 import { PHASE_PRODUCTION_BUILD } from 'next/dist/shared/lib/constants';
-import { Client } from 'pg';
 
+// import { Client } from 'pg';
 import * as schema from '@/models/Schema';
 
 import { Env } from './Env';
@@ -18,15 +20,20 @@ let drizzle;
 // Need a database for production? Check out https://www.prisma.io/?via=saasboilerplatesrc
 // Tested and compatible with Next.js Boilerplate
 if (process.env.NEXT_PHASE !== PHASE_PRODUCTION_BUILD && Env.DATABASE_URL) {
-  client = new Client({
+  client = new Pool({
     connectionString: Env.DATABASE_URL,
   });
-  await client.connect();
+  drizzle = drizzleNeon(client, { schema });
 
-  drizzle = drizzlePg(client, { schema });
-  await migratePg(drizzle, {
-    migrationsFolder: path.join(process.cwd(), 'migrations'),
-  });
+  // client = new Client({
+  //   connectionString: Env.DATABASE_URL,
+  // });
+  // await client.connect();
+  //
+  // drizzle = drizzlePg(client, { schema });
+  // await migratePg(drizzle, {
+  //   migrationsFolder: path.join(process.cwd(), 'migrations'),
+  // });
 } else {
   // Stores the db connection in the global scope to prevent multiple instances due to hot reloading with Next.js
   const global = globalThis as unknown as { client: PGlite; drizzle: PgliteDatabase<typeof schema> };